Assignment__GOOGLE__POM/src/main/java/pages/Google_03_AccountPage.java
package pages;

import org.openqa.selenium.WebElement;
import org.openqa.selenium.support.FindBy;
import org.openqa.selenium.support.PageFactory;

import base.TestBase;
import config.XPath;


public class Google_03_AccountPage extends TestBase {

	public Google_03_AccountPage() {
		PageFactory.initElements(driver, this);
	}

	@FindBy(xpath = "//a[@id='sdgBod']")
	public WebElement GoogleAccount;

	public boolean verifyGoogleAccount() {
		return GoogleAccount.isDisplayed();
	}
	
	@FindBy(xpath=XPath.Accountsingbtn)
	public WebElement Accountsingbtn;
	
	public Google_04_AccountVerifyPage clickonSigninButton(){
		Accountsingbtn.click();
		return new Google_04_AccountVerifyPage();
	}
	
	
	
	

}
